fix(async): Remove _exponentialBackoffWithJitter from public file
