+++ async/_util.ts
@@ -1,0 +1,14 @@
+// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+// This module is browser compatible.
+
+export function _exponentialBackoffWithJitter(
+    cap: number,
+    base: number,
+    attempt: number,
+    multiplier: number,
+    jitter: number,
+  ) {
+    const exp = Math.min(cap, base * multiplier ** attempt);
+    return (1 - jitter * Math.random()) * exp;
+  }
+
